cast added

+++ Exercises/WorkSheet3/src/adaptor/XpayToPayDAdapter.java
@@ -70,7 +70,7 @@ public class XpayToPayDAdapter implements PayD {
     private void adpat(){
         setCardOwnerName(this.xpay.getCustomerName());
         setCustCardNo(this.xpay.getCreditCardNo());
-        setCVVNo(this.xpay.getCardCVVNo());
+        setCVVNo((int) this.xpay.getCardCVVNo());
         setCardExpMonthDate(this.xpay.getCardExpMonth()+"/"+this.xpay.getCardExpYear());
         setTotalAmount(this.xpay.getAmount());
     }
